+++ src/package.ts
@@ -5,17 +5,19 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import * as path from 'path';
+import * as semver from 'semver';
+import { cli } from 'cli-ux';
 import { exec, pwd } from 'shelljs';
 import { fs, Logger, SfdxError } from '@salesforce/core';
-import { AsyncOptionalCreatable } from '@salesforce/kit';
-import { AnyJson, get } from '@salesforce/ts-types';
+import { AsyncOptionalCreatable, findKey } from '@salesforce/kit';
+import { AnyJson, get, Nullable } from '@salesforce/ts-types';
 import { Registry } from './registry';
 
 export type PackageJson = {
   name: string;
   version: string;
-  dependencies: AnyJson;
-  devDependencies: AnyJson;
+  dependencies: Record<string, string>;
+  devDependencies: Record<string, string>;
   scripts: Record<string, string>;
   files?: string[];
   pinnedDependencies?: string[];
@@ -52,6 +54,7 @@ interface PinnedPackage {
   name: string;
   version: string;
   tag: string;
+  alias: Nullable<string>;
 }
 
 export class Package extends AsyncOptionalCreatable {
@@ -127,11 +130,28 @@ export class Package extends AsyncOptionalCreatable {
         'Pinning package dependencies requires property "pinnedDependencies" to be present in package.json'
       );
     }
-    const dependencies: string[] = this.packageJson.pinnedDependencies;
+    const { pinnedDependencies, dependencies } = this.packageJson;
+    const deps = pinnedDependencies.map((d) => {
+      const version = dependencies[d];
+      if (version.startsWith('npm:')) {
+        return {
+          name: version.replace('npm:', '').replace(/@(\^|~)?[0-9]{1,3}(?:.[0-9]{1,3})?(?:.[0-9]{1,3})?(.*?)$/, ''),
+          version: version.split('@').reverse()[0].replace('^', '').replace('~', ''),
+          alias: d,
+        };
+      } else {
+        return {
+          name: d,
+          version: version.split('@').reverse()[0].replace('^', '').replace('~', ''),
+          alias: null,
+        };
+      }
+    });
+
     const pinnedPackages: PinnedPackage[] = [];
-    dependencies.forEach((name) => {
+    deps.forEach((dep) => {
       // get the 'release' tag version or the version specified by the passed in tag
-      const result = exec(`npm view ${name} dist-tags ${this.registry.getRegistryParameter()} --json`, {
+      const result = exec(`npm view ${dep.name} dist-tags ${this.registry.getRegistryParameter()} --json`, {
         silent: true,
       });
       const versions = JSON.parse(result.stdout) as Record<string, string>;
@@ -142,13 +162,28 @@ export class Package extends AsyncOptionalCreatable {
         tag = 'latest';
       }
 
-      const version = versions[tag];
+      // If the version in package.json is greater than the version of the requested tag, then we
+      // assume that this is on purpose - so we don't overwrite it. For example, we might want to
+      // include a latest-rc version for a single plugin but everything else we want latest.
+      let version: string;
+      if (semver.gt(dep.version, versions[tag])) {
+        cli.warn(
+          `${dep.name} is currently pinned at ${dep.version} which is higher than ${tag} (${versions[tag]}). Assuming that this is intentional...`
+        );
+        version = dep.version;
+        tag = findKey(versions, (v) => v === version);
+      } else {
+        version = versions[tag];
+      }
 
       // insert the new hardcoded versions into the dependencies in the project's package.json
-      this.packageJson['dependencies'][name] = version;
-
+      if (dep.alias) {
+        this.packageJson['dependencies'][dep.alias] = `npm:${dep.name}@${version}`;
+      } else {
+        this.packageJson['dependencies'][dep.name] = version;
+      }
       // accumulate information to return
-      pinnedPackages.push({ name, version, tag });
+      pinnedPackages.push({ name: dep.name, version, tag, alias: dep.alias });
     });
 
     return pinnedPackages;